pyproject.toml
[tool.poetry]
name = "gitscanner"
version = "0.1.0"
description = ""
authors = ["Fabian Becker <halfdan@xnorfz.de>"]

[tool.poetry.dependencies]
python = "^3.8"
GitPython = "^3.1.11"
PyGithub = "^1.54"
click = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
